+++ src/Args/Taxonomy.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Taxonomy class.
+ *
+ * @since 0.3.0
+ */
+
+namespace Required\Common\Args;
+
+use Args;
+
+/**
+ * Arguments for the `register_taxonomy()` function in WordPress.
+ *
+ * @since 0.3.0
+ */
+class Taxonomy extends Args\register_taxonomy {
+}